cloud/requirements.txt
amqp==1.4.9
anyjson==0.3.3
astroid==1.3.4
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.23
boto==2.27.0
celery==3.1.23
dj-database-url==0.3.0
Django==1.10.1
django-allauth==0.25.2
django-crispy-forms==1.6.0
django-debug-toolbar==1.4.0
django-loginas==0.1.9
django-mailgun==0.8.0
django-picklefield==0.3.2
django-qsstats-magic==0.7.2
django-queryset-csv==0.3.3
django-tables2==1.1.6
djangorestframework==3.3.3
envoy==0.0.3
Fabric==1.8.0
flower==0.7.0
humanize==0.5.1
itsdangerous==0.24
kombu==3.0.35
logilab-common==0.63.2
mock==1.0.1
oauthlib==0.7.2
paramiko==1.11.1
psycopg2==2.5.1
py-bcrypt==0.4
pycrypto==2.6
-e git+https://github.com/endaga/libpynexmo.git@d1dc03cbe0b8d68efffd3458c2b305cdf3267b0e#egg=pynexmo
python-dateutil==2.2
python-openid==2.2.5
pytz==2016.3
PyYAML==3.11
requests==2.0.0
requests-oauthlib==0.4.2
six==1.10.0
sqlparse==0.1.14
stripe==1.9.5
tornado==3.2.2
uWSGI==2.0.12
xlrd==0.9.3
django-guardian==1.4.2
snowflake
